refactor: 클라이언트 컴포넌트 도메인 계층으로 이동 및 MyPage CSS 재배치
- Client 전용 페이지 파일을 src/domains/user/pages 로 이동
- App Router 기반 라우트 구조에 맞게 MyPage CSS 모듈 정리
- 라우팅/스타일 파일 구조 일관성 개선
