+++ src/app/(user)/instructor/courses/page.tsx
@@ -1,4 +1,4 @@
-import InstructorClientpage from "./InstructorClientPage";
+import InstructorClientpage from "@/domains/user/pages/InstructorClientPage";
 
 export default async function InstructorCoursesPage() {
   return (
