+++ src/app/(user)/layout.tsx
@@ -1,4 +1,4 @@
-import layout from './MyPage.module.css';
+import layout from '@/app/(user)/mypage/MyPage.module.css';
 import MyPageSidebar from '@/domains/user/components/MyPageSidebar';
 import { AppShell } from '@/shared/ui/AppShell';
