+++ src/app/(user)/mypage/enrolled/page.tsx
@@ -1,4 +1,4 @@
-import EnrolledClientPage from "./EnrolledClientPage";
+import EnrolledClientPage from "@/domains/user/pages/EnrolledClientPage";
 
 export default async function EnrolledCoursesPage() {
   return (
