+++ src/app/(user)/mypage/page.tsx
@@ -1,4 +1,4 @@
-import Profile from "./profile/page";
+import Profile from "@/app/(user)/mypage/profile/page";
 
 export default function MyPage() {
   return <Profile />;
