+++ src/app/(user)/mypage/profile/page.tsx
@@ -1,4 +1,4 @@
-import ProfileClientPage from "./ProfileClientPage";
+import ProfileClientPage from "@/domains/user/pages/ProfileClientPage";
 
 export default async function ProfilePage() {
   return (
